requirements.txt
flask==2.3.3
pymupdf==1.24.3
python-dotenv==1.0.0
flask-cors==4.0.0